Create /mnt/userfs folder before doing the firmware upgrade.
This folder is needed for niinstallsafemode to run correctly on latest
OS versions on Arm targets. I don't delete it after the upgrade since the script will
clean any file it puts into that folder. And I don't want to remove the
folder since in theory can be created by the user and contain other
files.

This is a workaround since the already released versions of OS contain
the version of niisntallsafemode script that doesn't work on Arm
targets. RT Team said that this script should be designed to work on
run-mode as well (I will create a bug on their side to fix this particular issue)
so I am still discussing with them what are the implications and what
this commitment would mean.

Signed-off-by: Cristian Hotea <cristian.hotea@ni.com>